requirements.txt
Django==1.9.6
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
django-webpack-loader==0.3.0
httpie==0.9.3
